READ.ME
This course introduces a selection of Analytics Systems Technologies that are deployed in lab sessions throughout the analytics program. A multitude of Analytics Systems Technologies are used for different purposes to describe data numerically and graphically, for data visualization, file systems (HFS) for a large data mart, applications of structured query language, and filtering & transforming to ingest the data through scripting languages. Some of the tools are taught in greater detail (e.g. Python, Machine Learning), whereas others are introduced more broadly